backend/pyproject.toml
[tool.poetry]
name = "ocultar-backend"
version = "0.1.0"
description = ""
authors = ["grittypuffy <grittypuffy@riseup.net>"]
readme = "README.md"
packages = [{include = "ocultar_backend", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pyinstaller = {version = "^6.11.0", python = ">=3.12,<3.14"}
colorama = "^0.4.6"
fastapi = {version = "^0.115.2", extras = ["standard"] }
presidio-image-redactor = "^0.0.53"
presidio-analyzer = "^2.2.355"
presidio-anonymizer = "^2.2.355"
lingua-language-detector = "^2.0.2"
uvicorn = "^0.32.0"
pdf2image = "^1.17.0"
img2pdf = "^0.5.1"
pillow = "^11.0.0"
python-multipart = "^0.0.12"
spacy = {version = "^3.8.2", extras = ["en_core_web_lg"]}
pytesseract = "^0.3.13"
opencv-python = "^4.10.0.84"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
